fp prefix scan working for the simple combinational fp adder, there is an error in odd numbers which possibly is due to truncation.
changes made to the fp_adder is just removing the frac_diff shift as this takes place in the prefix_scan module at the input stage, and repeatedly performing the shift led to large error accumulation. This is something to look at later, it might need to be added back and the logic could be changed to support it.

test_prefix_scan is a python implementation of the algorithm used in the hardware, it compares the result of the algorithm with the actual cumsum. these results are used as the golden results in the testbench, and the testbench was restructured to be similar to that python program.